Refactor: Move `Session` into mod `session`
